Added router to main file, removed old similarity module
